Format query string to save to DB
