+++ utils/formatQueryTextToSave.ts
